Added new PilatusEnergy parameter.  This allows setting the camserver energy parameter independent of the threshold. Previously camserver was always setting the energy to 2*threshold.
When reading the camserver TIFF file the TIFFTAG_IMAGEDESCRIPTION field is read and added to the image as an
NDAttribute of type NDAttrString.  This allows capturing the important camserver settings as a long string.

Always read the detector status once in the constructor to get the TVX version, etc.

Add support for NDDriverVersion that was added in ADCore R2-6.